Make @mention dropdown not match when trailing space.
Otherwise you don't get rid of the dropdown when you select a user
who doesn't have a space in it. fixes #643
It then prevents tripping into a bug in textcomplete:
https://github.com/yuku-t/textcomplete/issues/115
(because when `@Joe ` is inserted without this patch then it matches again
and then the reinsertion if you select again will be an identical replacement
and then js hangs (haven't found out why).